pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "espanol-encuentro"
dynamic = ["version"]
description = 'Build your own Spanish dictionary'
readme = "README.md"
requires-python = ">=3.12"
authors = [
  { name = "Luke Chapman", email = "lrrchapman@gmail.com" },
]
dependencies = [
  "pydantic>=2.8",
  "pyyaml>=6.0",
  "click>=8.1",
]

[project.urls]
Documentation = "https://github.com/luke-chapman/espanol-encuentro#readme"
Issues = "https://github.com/luke-chapman/espanol-encuentro/issues"
Source = "https://github.com/luke-chapman/espanol-encuentro"

[project.scripts]
espanol-encuentro = "espanol_encuentro.__main__:main"
ee = "espanol_encuentro.__main__:main"

[tool.hatch.version]
path = "src/espanol_encuentro/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/espanol_encuentro tests}"
